django/pyproject.toml
[project]
name = "ksi-oidc-django"
description = "The OpenID Connect plugin Django created by KSIUJ"
requires-python = ">=3.13.0"
version = "0.8.0"
authors = [
    { name = "dominik-korsa" }
]
dependencies = [
    "django>=5.2.6",
    "django-solo>=2.4.0",
    "ksi-oidc-common>=0.5.1",
]
license="MPL-2.0"

[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"

[tool.ruff]
extend-exclude = ["src/ksi_oidc_django/migrations"]
